- added SSO endpoint "getAccessData()"
