+++ app/Api.php
@@ -1,0 +1,14 @@
+<?php
+/**
+ * Created by PhpStorm.
+ * User: Exodus 4D
+ * Date: 26.12.2018
+ * Time: 20:24
+ */
+
+namespace Exodus4D\ESI;
+
+
+abstract class Api implements ApiInterface {
+
+}
